refactor: move validation.js to utils folder
